[MINOR] - Renamed BeaconChainHeadHander and BeaconStateHander and moved into the new structure.
