+++ data/beaconrestapi/src/main/java/tech/pegasys/artemis/beaconrestapi/BeaconRestApi.java
@@ -24,11 +24,11 @@ import org.apache.commons.lang3.StringUtils;
 import tech.pegasys.artemis.api.ChainDataProvider;
 import tech.pegasys.artemis.api.DataProvider;
 import tech.pegasys.artemis.api.NetworkDataProvider;
-import tech.pegasys.artemis.beaconrestapi.beaconhandlers.BeaconChainHeadHandler;
-import tech.pegasys.artemis.beaconrestapi.beaconhandlers.BeaconStateHandler;
 import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetBlock;
+import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetChainHead;
 import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetCommittees;
 import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetHead;
+import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetState;
 import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetStateRoot;
 import tech.pegasys.artemis.beaconrestapi.handlers.beacon.GetValidators;
 import tech.pegasys.artemis.beaconrestapi.handlers.beacon.PostValidators;
@@ -115,10 +115,10 @@ public class BeaconRestApi {
   private void addBeaconHandlers(final DataProvider dataProvider) {
     final ChainDataProvider provider = dataProvider.getChainDataProvider();
     app.get(GetBlock.ROUTE, new GetBlock(provider, jsonProvider));
-    app.get(BeaconChainHeadHandler.ROUTE, new BeaconChainHeadHandler(provider, jsonProvider));
+    app.get(GetChainHead.ROUTE, new GetChainHead(provider, jsonProvider));
     app.get(GetHead.ROUTE, new GetHead(provider, jsonProvider));
     app.get(GetCommittees.ROUTE, new GetCommittees(provider, jsonProvider));
-    app.get(BeaconStateHandler.ROUTE, new BeaconStateHandler(provider, jsonProvider));
+    app.get(GetState.ROUTE, new GetState(provider, jsonProvider));
     app.get(GetStateRoot.ROUTE, new GetStateRoot(provider, jsonProvider));
 
     app.post(PostValidators.ROUTE, new PostValidators(provider, jsonProvider));
