+++ data/beaconrestapi/src/test/java/tech/pegasys/artemis/beaconrestapi/handlers/beacon/GetStateTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.beaconrestapi.beaconhandlers;
+package tech.pegasys.artemis.beaconrestapi.handlers.beacon;
 
 import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
 import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND;
@@ -44,7 +44,7 @@ import tech.pegasys.artemis.provider.JsonProvider;
 import tech.pegasys.artemis.storage.ChainStorageClient;
 import tech.pegasys.artemis.util.async.SafeFuture;
 
-public class BeaconStateHandlerTest {
+public class GetStateTest {
   private static tech.pegasys.artemis.datastructures.state.BeaconState beaconStateInternal;
   private static BeaconState beaconState;
   private static Bytes32 blockRoot;
@@ -68,7 +68,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnNotFoundWhenQueryAgainstMissingRootObject() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(ROOT, List.of(missingRoot)));
@@ -82,7 +82,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBadRequestWhenNoParameterSpecified() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(context.queryParamMap()).thenReturn(Collections.emptyMap());
 
@@ -93,7 +93,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBadRequestWhenBadSlotSpecified() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(SLOT, List.of("not-an-int")));
@@ -105,7 +105,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBadRequestWhenBadParamSpecified() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(EPOCH, List.of("not-an-int")));
@@ -117,7 +117,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBeaconStateObjectWhenQueryByRoot() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(ROOT, List.of(blockRoot.toHexString())));
@@ -135,7 +135,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBadRequestWhenEmptyRootIsSpecified() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(ROOT, List.of()));
@@ -147,7 +147,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBadRequestWhenEmptySlotIsSpecified() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(SLOT, List.of()));
@@ -159,7 +159,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBadRequestWhenMultipleParametersSpecified() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(SLOT, List.of(), ROOT, List.of()));
@@ -171,7 +171,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnBeaconStateObjectWhenQueryBySlot() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(SLOT, List.of(slot.toString())));
@@ -189,7 +189,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnNotFoundWhenQueryByMissingSlot() throws Exception {
-    BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(true);
     when(context.queryParamMap()).thenReturn(Map.of(SLOT, List.of("11223344")));
@@ -203,7 +203,7 @@ public class BeaconStateHandlerTest {
 
   @Test
   public void shouldReturnNoContentIfStoreNotDefined() throws Exception {
-    final BeaconStateHandler handler = new BeaconStateHandler(dataProvider, jsonProvider);
+    final GetState handler = new GetState(dataProvider, jsonProvider);
 
     when(dataProvider.isStoreAvailable()).thenReturn(false);
     when(context.queryParamMap()).thenReturn(Map.of(SLOT, List.of("11223344")));
